fix(pwrdrvr): align tarball contents with published baseline
